Verification Commit

+++ app/src/main/java/ru/skillbranch/devintensive/utils/Utils.kt
@@ -139,12 +139,12 @@ object Utils {
      */
     fun verification(github: String) = Regex("((https://|www.|https://www.)?github.com/(?!enterprise$|features$|topics$|collections$|trending$|events$|marketplace$|pricing$|nonprofit$|customer-stories$|security$|login$|join$)[\\w\\d-_]{1,39}$)|").find(github)?.value == github
 
-    fun dpToPx(dp: Float, context: Context): Int{
-        return TypedValue.applyDimension(TypedValue.COMPLEX_UNIT_DIP, dp, context.resources.displayMetrics).toInt()
+    fun dpToPx(dp: Int, context: Context): Int{
+        return TypedValue.applyDimension(TypedValue.COMPLEX_UNIT_DIP, dp.toFloat(), context.resources.displayMetrics).toInt()
     }
 
-    fun pxToDp(px: Float, context: Context): Int{
-        return (px / (context.resources.displayMetrics.densityDpi / DisplayMetrics.DENSITY_DEFAULT)).toInt()
+    fun pxToDp(px: Int, context: Context): Int{
+        return (px / (context.resources.displayMetrics.densityDpi / DisplayMetrics.DENSITY_DEFAULT))
     }
 }
